fix: move Steps component from Hero to Features section
