+++ seo-client/app/(landing)/_components/Features.tsx
@@ -1,10 +1,12 @@
 import CoreCapabilities from "./CoreCapabilities";
 import Testimonials from "./Testimonials";
 import EngineSection from "./EngineSection";
+import Steps from "./Steps";
 
 export default function Features() {
   return (
     <>
+      <Steps/>
       <CoreCapabilities />
       <Testimonials />
       <EngineSection />
